VideoCore: Split lighting regs from Regs struct
