+++ src/video_core/CMakeLists.txt
@@ -33,6 +33,7 @@ set(HEADERS
             rasterizer.h
             rasterizer_interface.h
             regs_framebuffer.h
+            regs_lighting.h
             regs_rasterizer.h
             regs_texturing.h
             renderer_base.h
